requirements.txt
absl-py==0.13.0
addict==2.4.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
# apex==0.1
apex
async-timeout==3.0.1
attrs==21.2.0
black==21.4b2
cachetools==4.2.2
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
configparser==5.0.2
contiguous-params==1.0.0
cython==0.29.23
dataclasses==0.6
decorator==4.4.2
dill==0.3.4
docker-pycreds==0.4.0
dotmap==1.3.23
einops==0.3.0
fairseq==0.10.2
fire==0.4.0
fsspec==2021.6.1
ftfy==6.0.3
future==0.18.2
# fvcore==0.1.5
fvcore
gitdb==4.0.7
gitpython==3.1.18
google-auth==1.32.1
google-auth-oauthlib==0.4.4
grpcio==1.38.1
hydra-core==1.1.0
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-resources==5.2.0
iopath==0.1.8
jpeg4py==0.1.4
lightseq==2.0.3
lmdb==1.2.1
markdown==3.3.4
matplotlib==3.3.4
mmcv-full==1.3.9
model-index==0.1.11
moviepy==1.0.3
msgpack-python==0.5.6
multidict==5.1.0
multiprocess==0.70.12.2
mypy-extensions==0.4.3
ninja==1.10.0.post2
numpy==1.20.3
# nvidia-dali-cuda110==1.3.0
nvidia-dali-cuda110
nvidia-pyindex==1.0.9
oauthlib==3.1.1
omegaconf==2.1.0
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
openmim==0.1.2
ordered-set==4.0.2
pandas==1.2.3
parameterized==0.8.1
pathos==0.2.8
pathspec==0.9.0
pathtools==0.1.2
pillow==8.2.0
pip==21.0.1
portalocker==2.0.0
pox==0.3.0
ppft==1.6.6.4
prefetch-generator==1.0.1
proglog==0.1.9
promise==2.3
protobuf==3.15.6
psutil==5.8.0
pyarrow==4.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pydeprecate==0.3.0
pydot==1.4.2
pytorch-lightning==1.3.8
pytorchvideo==0.1.2
pytz==2021.1
pyyaml==5.4.1
regex==2021.7.1
requests-oauthlib==1.3.0
rsa==4.7.2
sacrebleu==1.5.1
scipy==1.6.1
sentry-sdk==1.3.1
shortuuid==1.0.1
simplejson==3.17.3
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardx==2.1
termcolor==1.1.0
toml==0.10.2

typing-extensions==3.7.4.3
umsgpack==0.1.0
wandb==0.15.3
wcwidth==0.2.5
werkzeug==2.0.1
yacs==0.1.8
yapf==0.31.0
yarl==1.6.3
zipp==3.4.1
